Delete 구현 중..

+++ WebApp/src/main/java/com/lhw/jins/album/AlbumDAO.java
@@ -1,10 +1,13 @@
 package com.lhw.jins.album;
 
+import static org.hamcrest.CoreMatchers.nullValue;
+
 import java.net.URLEncoder;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
+import org.apache.ibatis.jdbc.Null;
 import org.apache.ibatis.session.SqlSession;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -21,7 +24,8 @@ public class AlbumDAO {
 	public void insert(Album album, HttpServletRequest request, HttpServletResponse response) {
 		try {
 			String path = request.getSession().getServletContext().getRealPath("resources/album/images/fulls");
-			MultipartRequest mr = new MultipartRequest(request, path, 31457280,	"euc-kr", new DefaultFileRenamePolicy());
+			MultipartRequest mr = new MultipartRequest(request, path, 31457280, "euc-kr",
+					new DefaultFileRenamePolicy());
 
 			String album_img = mr.getFilesystemName("album_img");
 			album_img = URLEncoder.encode(album_img, "euc-kr");
@@ -37,7 +41,7 @@ public class AlbumDAO {
 				System.out.println("사진 INSERT 실패");
 			}
 		} catch (Exception e) {
-			System.out.println("사진 INSERT 실패");
+			System.out.println("사진 INSERT 실패(DB 문제)");
 			e.printStackTrace();
 		} finally {
 		}
@@ -57,7 +61,7 @@ public class AlbumDAO {
 			album.setAlbum_img(jins_img);
 			album.setAlbum_txt1(mr.getParameter("album_txt1"));
 			album.setAlbum_txt2(mr.getParameter("album_txt2"));
-			
+
 			if (ss.getMapper(AlbumMapper.class).updateAlbum(album) == 1) {
 				System.out.println("사진 수정 성공");
 			} else {
@@ -68,12 +72,22 @@ public class AlbumDAO {
 			System.out.println("DB서버 문제");
 		}
 	}
-	
+
 	public void getAllAlbum(HttpServletRequest request, HttpServletResponse response) {
 		request.setAttribute("albumList", ss.getMapper(AlbumMapper.class).selectAlbum());
 	}
-	
+
 	public void delete(Album album, HttpServletRequest request, HttpServletResponse response) {
-		request.setAttribute("albumList", ss.getMapper(AlbumMapper.class).deleteAlbum(album));
+		try {
+			if (ss.getMapper(AlbumMapper.class).deleteAlbum(album) == 0) {
+				System.out.println("없는 사진 삭제");
+			} else {
+				request.setAttribute("albumList", ss.getMapper(AlbumMapper.class).deleteAlbum(album));
+				System.out.println("사진 삭제 성공");
+			}
+		} catch (Exception e) {
+			e.printStackTrace();
+			System.out.println("DB 문제");
+		}
 	}
 }
